Prototype: Implementation of CountedCompleter (ForkJoinTask). Second version. Simple prototype of CountedCompleter class for recursive computation into ForkJoinPool.
